+++ MyPay/Model/User.swift
@@ -21,7 +21,7 @@ public class User {
     public var securityCodeSalt: String = ""
     
     // id (initialized right after user's data gets inserted into the database)
-    private var userId: Int = -1
+    private var userId: Int16 = -1
     
     public init(_ firstName: String, _ lastName: String, _ birthDate: String, _ areaCode: String, _ phoneNumber: String) {
         self.firstName = firstName
